requirements.txt
google-generativeai
python-dotenv
ytmusicapi
streamlit